fix(deps): leverage the github workflows predicate
and convert semantic-release lifter tests to vitest

for #458
