+++ package.json
@@ -63,7 +63,7 @@
   "dependencies": {
     "@form8ion/config-file": "^1.0.1",
     "@form8ion/core": "^4.6.1",
-    "@form8ion/github-workflows-core": "^5.4.0",
+    "@form8ion/github-workflows-core": "^5.5.0",
     "@form8ion/javascript-core": "^11.0.0",
     "deepmerge": "^4.2.2"
   },
