--- src/semantic-release/ci-providers/github-workflows/tester.js
@@ -1,5 +1,0 @@
-import {directoryExists} from '@form8ion/core';
-
-export default function ({projectRoot}) {
-  return directoryExists(`${projectRoot}/.github/workflows`);
-}
